[#12] 스프라이트 x, y좌표 Min 미만 짤림 구현
- 스프라이트 짤림 구현 완료!

+++ src/render/sprite.cpp
@@ -8,7 +8,7 @@ void Sprite::start_next_frame()
 	frame_cnt = frame_cnt > 59 ? 0 : frame_cnt + 1;
 }
 
-Sprite::Sprite(byte _face_id, byte _x, byte _y)
+Sprite::Sprite(byte _face_id, short _x, short _y)
 {
 	x = _x;
 	y = _y;
