+++ src/render/sprite.hpp
@@ -6,14 +6,14 @@
 class Sprite
 {
 public:
-	Sprite(byte _face_id, byte _x, byte _y);
+	Sprite(byte _face_id, short _x, short _y);
 	static byte frame_cnt;
 	static void start_next_frame();
 
-	byte x;
-	byte y;
-	byte vx;
-	byte vy;
+	short x;
+	short y;
+	int8_t vx;
+	int8_t vy;
 	byte x_frame;
 	byte y_frame;
 	byte face_id;
